+++ frontend/src/scenes/automations/AutomationStepSidebar/AutomationStepConfig.scss
