+++ frontend/src/scenes/automations/AutomationStepSidebar/automationStepMenuLogic.ts
@@ -1,6 +1,12 @@
 import { actions, kea, path, reducers, selectors } from 'kea'
-import type { automationStepConfigLogicType } from './automationStepConfigLogicType'
-import { AnyAutomationStep, AutomationStepCategory, AutomationStepConfigType, AutomationStepKind } from './schema'
+import type { automationStepConfigLogicType } from '../automationStepConfigLogicType'
+import { AnyAutomationStep, AutomationStepCategory, AutomationStepConfigType, AutomationStepKind } from '../schema'
+
+const DEFAULT_EVENT: ActionFilter = {
+    id: '$pageview',
+    math: 'dau',
+    type: 'events',
+}
 
 import {
     // GithubIcon,
@@ -17,6 +23,7 @@ import {
     IconWebhook,
 } from 'lib/lemon-ui/icons'
 import { EventSentConfig } from './AutomationStepConfig'
+import { ActionFilter } from '~/types'
 import { uuid } from 'lib/utils'
 
 const stepOptions: AnyAutomationStep[] = [
@@ -24,7 +31,7 @@ const stepOptions: AnyAutomationStep[] = [
         kind: AutomationStepKind.EventSource,
         id: 'new',
         category: AutomationStepCategory.Source,
-        filters: [],
+        filters: [DEFAULT_EVENT],
     },
     // { kind: AutomationStepKind.ActionSource, category: AutomationStepCategory.Source },
     // { kind: AutomationStepKind.PauseForLogic, category: AutomationStepCategory.Logic },
@@ -94,7 +101,7 @@ export const automationStepConfigLogic = kea<automationStepConfigLogicType>([
                     kind: AutomationStepKind.EventSource,
                     id: uuid(),
                     category: AutomationStepCategory.Source,
-                    filters: [],
+                    filters: [DEFAULT_EVENT],
                 },
             ] as AnyAutomationStep[],
             {
